+++ lib/src/core/better_player_controller.dart
@@ -121,6 +121,9 @@ class BetterPlayerController extends ChangeNotifier {
   Stream<bool> get controlsVisibilityStream =>
       _controlsVisibilityStreamController.stream;
 
+  AppLifecycleState _appLifecycleState = AppLifecycleState.resumed;
+  BetterPlayerEventType _betterPlayerEventBeforePause;
+
   BetterPlayerController(
     this.betterPlayerConfiguration, {
     this.betterPlayerPlaylistConfiguration,
@@ -313,9 +316,12 @@ class BetterPlayerController extends ChangeNotifier {
   }
 
   Future<void> play() async {
-    await videoPlayerController.play();
-    _hasCurrentDataSourceStarted = true;
-    _postEvent(BetterPlayerEvent(BetterPlayerEventType.PLAY));
+    _betterPlayerEventBeforePause = BetterPlayerEventType.PLAY;
+    if (_appLifecycleState == AppLifecycleState.resumed) {
+      await videoPlayerController.play();
+      _hasCurrentDataSourceStarted = true;
+      _postEvent(BetterPlayerEvent(BetterPlayerEventType.PLAY));
+    }
   }
 
   Future<void> setLooping(bool looping) async {
@@ -323,6 +329,7 @@ class BetterPlayerController extends ChangeNotifier {
   }
 
   Future<void> pause() async {
+    _betterPlayerEventBeforePause = BetterPlayerEventType.PAUSE;
     await videoPlayerController.pause();
     _postEvent(BetterPlayerEvent(BetterPlayerEventType.PAUSE));
   }
@@ -552,6 +559,15 @@ class BetterPlayerController extends ChangeNotifier {
   }
 
   bool get hasCurrentDataSourceStarted => _hasCurrentDataSourceStarted;
+
+  void setAppLifecycleState(AppLifecycleState appLifecycleState) {
+    _appLifecycleState = appLifecycleState;
+    if (appLifecycleState == AppLifecycleState.resumed) {
+      if (_betterPlayerEventBeforePause == BetterPlayerEventType.PLAY) {
+        play();
+      }
+    }
+  }
 
   @override
   void dispose() {